app/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
flask==3.0.3
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
package-name==0.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
gunicorn
boto3
prometheus-client
